dev-lang/luajit: Use luajit_relver in bin symlink
Previously we were generating this symlink:

    /usr/bin/luajit -> /usr/bin/luajit-${PV}

This failed for 2.1.9999999999, as the binary was instead at
/usr/bin/luajit-2.1.${relver}, for the corresponding relver
of the commit.

We now read `src/luajit_relver.txt` generated by luajit's
build process, which always contains the correct relver
suffix for both snapshot and live builds.

The 9999999999 support is also now agnostic to which release
channel is being built, so a theoretical 2.2.9999999999
could be supported by copying the 2.1.9999999999 ebuild
without modification.

Closes: https://bugs.gentoo.org/943404
Signed-off-by: Artemis Everfree <artemis@artemis.sh>